feat: projection type 수정

+++ backend/src/main/java/com/imsnacks/Nyeoreumnagi/weather/service/projection_entity/WindInfo.java
@@ -1,6 +1,6 @@
 package com.imsnacks.Nyeoreumnagi.weather.service.projection_entity;
 
 public interface WindInfo {
-    Double getMaxWindSpeed();
-    Double getWindDirection();
+    Integer getMaxWindSpeed();
+    Integer getWindDirection();
 }
